listening-comp/requirements.txt
boto3
youtube_transcript_api
streamlit>=1.24.0
flask>=2.0.1
flask-cors>=3.0.10
numpy>=1.21.0
mediapipe>=0.10.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
pillow>=9.5.0
requests>=2.26.0
python-dotenv>=0.19.0